refactor(Board): DtoMapper를 이용한 DTO 변환 구현
