+++ src/main/java/page/clab/api/domain/community/board/application/dto/request/BoardRequestDto.java
@@ -4,10 +4,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.Setter;
-import page.clab.api.domain.community.board.domain.Board;
 import page.clab.api.domain.community.board.domain.BoardCategory;
-import page.clab.api.global.common.file.domain.UploadedFile;
-import page.clab.api.global.util.RandomNicknameUtil;
 
 import java.util.List;
 
@@ -36,18 +33,4 @@ public class BoardRequestDto {
     @NotNull(message = "{notNull.board.wantAnonymous}")
     @Schema(description = "익명 사용 여부", example = "false", required = true)
     private boolean wantAnonymous;
-
-    public static Board toEntity(BoardRequestDto requestDto, String memberId, List<UploadedFile> uploadedFiles) {
-        return Board.builder()
-                .memberId(memberId)
-                .nickname(RandomNicknameUtil.makeRandomNickname())
-                .category(requestDto.getCategory())
-                .title(requestDto.getTitle())
-                .content(requestDto.getContent())
-                .uploadedFiles(uploadedFiles)
-                .imageUrl(requestDto.getImageUrl())
-                .wantAnonymous(requestDto.isWantAnonymous())
-                .isDeleted(false)
-                .build();
-    }
 }
