+++ src/main/java/page/clab/api/domain/community/board/application/dto/response/BoardCategoryResponseDto.java
@@ -2,8 +2,6 @@ package page.clab.api.domain.community.board.application.dto.response;
 
 import lombok.Builder;
 import lombok.Getter;
-import page.clab.api.domain.community.board.domain.Board;
-import page.clab.api.domain.memberManagement.member.application.dto.shared.MemberDetailedInfoDto;
 
 import java.time.LocalDateTime;
 
@@ -19,18 +17,4 @@ public class BoardCategoryResponseDto {
     private Long commentCount;
     private String imageUrl;
     private LocalDateTime createdAt;
-
-    public static BoardCategoryResponseDto toDto(Board board, MemberDetailedInfoDto memberInfo, Long commentCount) {
-        WriterInfo writerInfo = WriterInfo.fromBoard(board, memberInfo);
-        return BoardCategoryResponseDto.builder()
-                .id(board.getId())
-                .category(board.getCategory().getKey())
-                .writerId(writerInfo.getId())
-                .writerName(writerInfo.getName())
-                .title(board.getTitle())
-                .commentCount(commentCount)
-                .imageUrl(board.getImageUrl())
-                .createdAt(board.getCreatedAt())
-                .build();
-    }
 }
