+++ src/main/java/page/clab/api/domain/community/board/application/service/BoardDetailsRetrievalService.java
@@ -4,6 +4,7 @@ import jakarta.persistence.Tuple;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import page.clab.api.domain.community.board.application.dto.mapper.BoardDtoMapper;
 import page.clab.api.domain.community.board.application.dto.response.BoardDetailsResponseDto;
 import page.clab.api.domain.community.board.application.dto.response.BoardEmojiCountResponseDto;
 import page.clab.api.domain.community.board.application.port.in.RetrieveBoardDetailsUseCase;
@@ -32,7 +33,7 @@ public class BoardDetailsRetrievalService implements RetrieveBoardDetailsUseCase
         MemberDetailedInfoDto memberInfo = externalRetrieveMemberUseCase.getMemberDetailedInfoById(board.getMemberId());
         boolean isOwner = board.isOwner(currentMemberInfo.getMemberId());
         List<BoardEmojiCountResponseDto> emojiInfos = getBoardEmojiCountResponseDtoList(boardId, currentMemberInfo.getMemberId());
-        return BoardDetailsResponseDto.toDto(board, memberInfo, isOwner, emojiInfos);
+        return BoardDtoMapper.toBoardDetailsResponseDto(board, memberInfo, isOwner, emojiInfos);
     }
 
     @Transactional(readOnly = true)
