+++ src/main/java/page/clab/api/domain/community/board/application/service/BoardRegisterService.java
@@ -3,6 +3,7 @@ package page.clab.api.domain.community.board.application.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import page.clab.api.domain.community.board.application.dto.mapper.BoardDtoMapper;
 import page.clab.api.domain.community.board.application.dto.request.BoardRequestDto;
 import page.clab.api.domain.community.board.application.port.in.RegisterBoardUseCase;
 import page.clab.api.domain.community.board.application.port.out.RegisterBoardPort;
@@ -33,7 +34,7 @@ public class BoardRegisterService implements RegisterBoardUseCase {
     public String registerBoard(BoardRequestDto requestDto) throws PermissionDeniedException {
         MemberDetailedInfoDto currentMemberInfo = externalRetrieveMemberUseCase.getCurrentMemberDetailedInfo();
         List<UploadedFile> uploadedFiles = uploadedFileService.getUploadedFilesByUrls(requestDto.getFileUrlList());
-        Board board = BoardRequestDto.toEntity(requestDto, currentMemberInfo.getMemberId(), uploadedFiles);
+        Board board = BoardDtoMapper.toBoard(requestDto, currentMemberInfo.getMemberId(), uploadedFiles);
         board.validateAccessPermissionForCreation(currentMemberInfo);
         if (board.shouldNotifyForNewBoard(currentMemberInfo)) {
             externalSendNotificationUseCase.sendNotificationToMember(currentMemberInfo.getMemberId(), "[" + board.getTitle() + "] 새로운 공지사항이 등록되었습니다.");
