+++ src/main/java/page/clab/api/domain/community/board/application/service/BoardsByCategoryRetrievalService.java
@@ -5,6 +5,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import page.clab.api.domain.community.board.application.dto.mapper.BoardDtoMapper;
 import page.clab.api.domain.community.board.application.dto.response.BoardCategoryResponseDto;
 import page.clab.api.domain.community.board.application.port.in.RetrieveBoardsByCategoryUseCase;
 import page.clab.api.domain.community.board.application.port.out.RetrieveBoardPort;
@@ -29,7 +30,7 @@ public class BoardsByCategoryRetrievalService implements RetrieveBoardsByCategor
         Page<Board> boards = retrieveBoardPort.findAllByCategory(category, pageable);
         return new PagedResponseDto<>(boards.map(board -> {
             long commentCount = externalRetrieveCommentUseCase.countByBoardId(board.getId());
-            return  BoardCategoryResponseDto.toDto(board, getMemberDetailedInfoByBoard(board), commentCount);
+            return  BoardDtoMapper.toBoardCategoryResponseDto(board, getMemberDetailedInfoByBoard(board), commentCount);
         }));
     }
