+++ src/main/java/page/clab/api/domain/community/board/application/service/DeletedBoardsRetrievalService.java
@@ -6,6 +6,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import page.clab.api.domain.community.board.application.dto.mapper.BoardDtoMapper;
 import page.clab.api.domain.community.board.application.dto.response.BoardListResponseDto;
 import page.clab.api.domain.community.board.application.port.in.RetrieveDeletedBoardsUseCase;
 import page.clab.api.domain.community.board.application.port.out.RetrieveBoardPort;
@@ -38,6 +39,6 @@ public class DeletedBoardsRetrievalService implements RetrieveDeletedBoardsUseCa
     @NotNull
     private BoardListResponseDto mapToBoardListResponseDto(Board board, MemberDetailedInfoDto memberInfo) {
         Long commentCount = externalRetrieveCommentUseCase.countByBoardId(board.getId());
-        return BoardListResponseDto.toDto(board, memberInfo, commentCount);
+        return BoardDtoMapper.toBoardListResponseDto(board, memberInfo, commentCount);
     }
 }
