+++ src/main/java/page/clab/api/domain/community/board/application/service/MyBoardsRetrievalService.java
@@ -5,6 +5,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import page.clab.api.domain.community.board.application.dto.mapper.BoardDtoMapper;
 import page.clab.api.domain.community.board.application.dto.response.BoardMyResponseDto;
 import page.clab.api.domain.community.board.application.port.in.RetrieveMyBoardsUseCase;
 import page.clab.api.domain.community.board.application.port.out.RetrieveBoardPort;
@@ -25,6 +26,6 @@ public class MyBoardsRetrievalService implements RetrieveMyBoardsUseCase {
     public PagedResponseDto<BoardMyResponseDto> retrieveMyBoards(Pageable pageable) {
         MemberBasicInfoDto currentMemberInfo = externalRetrieveMemberUseCase.getCurrentMemberBasicInfo();
         Page<Board> boards = retrieveBoardPort.findAllByMemberId(currentMemberInfo.getMemberId(), pageable);
-        return new PagedResponseDto<>(boards.map(board -> BoardMyResponseDto.toDto(board, currentMemberInfo)));
+        return new PagedResponseDto<>(boards.map(board -> BoardDtoMapper.toBoardMyResponseDto(board, currentMemberInfo)));
     }
 }
